Middleware and utils package were created and some fuctions were moved to those packages.
