+++ internal/api/handlers/login.go
@@ -1,22 +1,17 @@
 package handlers
 
 import (
+	"RestAPIGolang/internal/auth"
 	"RestAPIGolang/internal/database"
 	model "RestAPIGolang/internal/models"
+	"RestAPIGolang/internal/utils"
 	"encoding/json"
 	"fmt"
 	"net/http"
-	"os"
-
-	"github.com/golang-jwt/jwt/v4"
 )
 
 func Login(w http.ResponseWriter, r *http.Request) error {
-	if r.Method != "POST" {
-		return fmt.Errorf("method not allowed %s", r.Method)
-	}
-
-	var req model.LoginRequest //add model
+	var req model.LoginRequest 
 	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
 		return err
 	}
@@ -30,7 +25,7 @@ func Login(w http.ResponseWriter, r *http.Request) error {
 		return fmt.Errorf("not authenticated")
 	}
 
-	token, err := createJWT(acc)
+	token, err := auth.CreateJWT(acc)
 	if err != nil {
 		return err
 	}
@@ -40,78 +35,5 @@ func Login(w http.ResponseWriter, r *http.Request) error {
 		Number: acc.Number,
 	}
 
-	return WriteJSON(w, http.StatusOK, resp)
-}
-
-// add modelo
-func createJWT(account *model.Account) (string, error) {
-	claims := &jwt.MapClaims{
-		"expiresAt":     15000,
-		"accountNumber": account.Number,
-	}
-
-	secret := os.Getenv("JWT_SECRET")
-	token := jwt.NewWithClaims(jwt.SigningMethodHS256, claims)
-
-	return token.SignedString([]byte(secret))
-}
-
-func permissionDenied(w http.ResponseWriter) {
-	WriteJSON(w, http.StatusForbidden, ApiError{Error: "permission denied"})
-}
-
-// eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJhY2NvdW50TnVtYmVyIjo0OTgwODEsImV4cGlyZXNBdCI6MTUwMDB9.TdQ907o9yhUI2KU0TngrqO-xbfNgHAfZI6Jngia15UE
-
-func WithJWTAuth(handlerFunc http.HandlerFunc) http.HandlerFunc {
-	return func(w http.ResponseWriter, r *http.Request) {
-		fmt.Println("calling JWT auth middleware")
-
-		tokenString := r.Header.Get("x-jwt-token")
-		token, err := validateJWT(tokenString)
-		if err != nil {
-			permissionDenied(w)
-			return
-		}
-		if !token.Valid {
-			permissionDenied(w)
-			return
-		}
-		userID, err := getID(r)
-		if err != nil {
-			permissionDenied(w)
-			return
-		}
-		account, err := database.GetAccountByID(userID)
-		if err != nil {
-			permissionDenied(w)
-			return
-		}
-
-		claims := token.Claims.(jwt.MapClaims)
-		if account.Number != int64(claims["accountNumber"].(float64)) {
-			permissionDenied(w)
-			return
-		}
-
-		if err != nil {
-			WriteJSON(w, http.StatusForbidden, ApiError{Error: "invalid token"})
-			return
-		}
-
-		handlerFunc(w, r)
-	}
-}
-
-func validateJWT(tokenString string) (*jwt.Token, error) {
-	secret := os.Getenv("JWT_SECRET")
-
-	return jwt.Parse(tokenString, func(token *jwt.Token) (interface{}, error) {
-		// Don't forget to validate the alg is what you expect:
-		if _, ok := token.Method.(*jwt.SigningMethodHMAC); !ok {
-			return nil, fmt.Errorf("unexpected signing method: %v", token.Header["alg"])
-		}
-
-		// hmacSampleSecret is a []byte containing your secret, e.g. []byte("my_secret_key")
-		return []byte(secret), nil
-	})
+	return utils.WriteJSON(w, http.StatusOK, resp)
 }
