+++ internal/api/middlewares/httpHandler.go
@@ -1,0 +1,19 @@
+package middlewares
+
+import (
+	"RestAPIGolang/internal/utils"
+	"net/http"
+)
+
+type ApiFunc func(http.ResponseWriter, *http.Request) error
+type ApiError struct {
+	Error string `json:"error"`
+}
+
+func MakeHTTPHandleFunc(f ApiFunc) http.HandlerFunc {
+	return func(w http.ResponseWriter, r *http.Request) {
+		if err:=f(w,r); err!=nil{
+			utils.WriteJSON(w, http.StatusBadRequest, ApiError{Error: err.Error()})
+		}
+	}
+}
