+++ internal/models/general.go
@@ -2,24 +2,7 @@ package model
 
 import (
 	"database/sql"
-	"time"
-
-	"golang.org/x/crypto/bcrypt"
 )
-
-type Account struct {
-	ID                int       `json:"id"`
-	FirstName         string    `json:"firstName"`
-	LastName          string    `json:"lastName"`
-	Number            int64     `json:"number"`
-	EncryptedPassword string    `json:"-"`
-	Balance           int64     `json:"balance"`
-	CreatedAt         time.Time `json:"createdAt"`
-}
-func (a *Account) ValidPassword(pw string) bool {
-	return bcrypt.CompareHashAndPassword([]byte(a.EncryptedPassword), []byte(pw)) == nil
-}
-
 
 type PostgresStore struct {
 	DB *sql.DB
