+++ internal/utils/seeding.go
@@ -1,15 +1,25 @@
-package helpers
+package utils
 
 import (
-	"RestAPIGolang/internal/api/handlers"
 	"RestAPIGolang/internal/database"
 	model "RestAPIGolang/internal/models"
+	"encoding/json"
 	"fmt"
 	"log"
+	"net/http"
 )
 
+
+func WriteJSON(w http.ResponseWriter, status int, v any) error {
+	w.Header().Add("Content-Type", "application/json")
+	w.WriteHeader(status)
+	return json.NewEncoder(w).Encode(v)
+}
+
+
 func seedAccount(fname, lname, pw string) *model.Account {
-	acc, err := handlers.NewAccount(fname, lname, pw)
+	acc:=new(model.Account)
+	err := acc.NewAccount(fname, lname, pw)
 	if err != nil {
 		log.Fatal(err)
 	}
@@ -26,3 +36,4 @@ func seedAccount(fname, lname, pw string) *model.Account {
 func SeedAccounts() {
 	seedAccount("anthony", "GG", "hunter88888")
 }
+
